requirements.txt
# Web / API
flask>=2.3.0
gunicorn>=21.2.0

# LangGraph agentic pipeline
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# LLM
requests>=2.31.0

# PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# RDF / SHACL
rdflib>=7.0.0
pyshacl>=0.25.0

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
